requirements.txt
Flask==2.0.0
Flask-MySQL==1.5.2
Jinja2==3.0.0
SQLAlchemy==1.3.20
Werkzeug==2.0.0
pandas==1.2.4
matplotlib==3.4.2
mysql-connector-python==8.0.25
numpy==1.20.3
